Fix --uppercase=false not working
Closes #16
Fixes the value of the uppercase option always being true.
This was due to the design of claps boolean flags which is
discussed here:
https://github.com/clap-rs/clap/issues/1649#issuecomment-1837123432
This PR uses the solution mentioned in the linked comment from
that thread.

Using `require_equals` and `default_missing_value` allows
`sleek --uppercase "queries/*sql"` to still work without
trying to parse the path as the uppercase boolean, reducing
breakage.

Use `--uppercase=false` to disable uppercase, NOT `--uppercase false`.